requirements.txt
requests
RPi.GPIO
waveshare-epaper
Pillow
